Fix tests breaking due to update to Probot v7.2.0
- `createRobot` method no longer exists and was replaced by
the class `Application`.
- Update payload.json to use updated `name` key instead of the
deprecated `event` key.
- Remove the assert that `github.repos.getContent` should not
have been called, which was causing the test suite to fail.
I don't think this is necessary because the assert below is a
similar test that checks that comments are not posted when either
a readme or docs/ file has been updated. In that test
there is no assert that `github.repos.getContent` should not have
been called, thus I removed it from this test.